@@ -2,7 +2,7 @@
     "name": "silktide/semrush-api",
     "description": "SEMrush API client",
     "require": {
-        "php": ">=7.1",
+        "php": ">=7.4",
         "guzzlehttp/guzzle": "^6.1 | ^7.0",
         "psr/log": "^1.0",
         "psr/simple-cache": "^1.0",
